+++ changelog.md
@@ -1,7 +1,12 @@
 # Change Log
 
 ## WIP
+
 - Drop support for GHC older than 8.10
+- Safe construction of vectors from linked lists
+  - https://github.com/expipiplus1/vector-sized/pull/88
+
+Thanks to @sheaf and @kozross
 
 ## [1.5.0] - 2021-08-25